Move FIDO2 public API to fido2_common.go, introduce IsFIDO2Available
